chore: separate shank render integration tests
