+++ shank-render/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod consts;
 mod render_seeds;
+pub use render_seeds::*;
